🔨 refactor: 불필요한 코드 정리
